requirements.txt
langchain
openai
langchain-openai
python-dotenv
huggingface_hub
streamlit